pay_group/requirements.txt
Flask
qrcode
wechatpy
pyzbar
Pillow
werkzeug
